server: remove abstraction of App
